p2p2-main/requirements.txt
flask==2.2.3
flask-login==0.6.2
flask-sqlalchemy==3.0.3
werkzeug==2.2.3
flask-socketio==5.3.4
web3>=7.0.0
bip44>=0.1.5
eth-account>=0.13.0
python-dotenv
gunicorn
cryptography
